+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.8.2 (November 18, 2022)
+
+ENHANCEMENTS:
+* Multiple improvements in ONTAP REST API client based on OnTap v9.12
+  * Moved LunCopy routine from PrivateCli /storage/luns REST API
+
 ## 1.8.1 (October 18, 2022)
 
 BUG FIXES: